gfx: widen mesh layout_hash to 64 bits
The mesh pipeline cache treats its key as identity — entries are matched by
key equality alone, descriptors are never compared on a hit. Every other
ingredient of that key is already 64-bit (resolved_vs/fs, render_state_hash),
but layout_hash was 32-bit: a leftover from the original key, which packed
`(layout_hash << 32) | program_handle` and sized the hash to fit beside a
32-bit handle. The packing is long gone; the width stayed.

Two layouts colliding in 32 bits would silently share a pipeline and bind the
wrong attribute setup. The hashed population is distinct vertex layouts —
tens of values in a real game, so the practical risk was already negligible.
This is an alignment with the rest of the key, not a bug fix: the hash is
computed once at mesh activation, so widening it costs no per-frame work and
8 bytes per mesh info slot.

Document the cache contract while here — a 64-bit hash serving as identity
with no descriptor comparison is a deliberate choice that reads like an
oversight, and it carries a maintenance rule: every nt_pipeline_desc_t field
a renderer varies has to be folded into its key.

+++ engine/graphics/nt_gfx.h
@@ -77,7 +77,7 @@ typedef struct {
     uint8_t index_type;                        /* 0=none, 1=uint16, 2=uint32 */
     NtStreamDesc streams[NT_MESH_MAX_STREAMS]; /* copied from pack data at activation */
     uint16_t stride;                           /* total vertex size in bytes */
-    uint32_t layout_hash;                      /* hash of stream descriptors for pipeline cache key */
+    uint64_t layout_hash;                      /* stream-descriptor hash; serves as pipeline-cache identity, hence 64-bit */
 } nt_gfx_mesh_info_t;
 
 /* ---- Enums ---- */
